Separate out typescript code into its own files
- Introduce dom.ts, event.ts, iso.ts, util.ts
