+++ js/miso.js
@@ -1,1 +1,1 @@
-function X(q,z,H){if(!q&&!z)return;else if(!q&&z)d(z,H);else if(!z)k(q,H);else if(q.type===z.type)u(q,z,H);else O(q,z,H)}var O=function(q,z,H){if(E(q),z.type==="vtext")z.domRef=document.createTextNode(z.text),H.replaceChild(z.domRef,q.domRef);else D(z,function(G){H.replaceChild(G,q.domRef)});R(q)},k=function(q,z){E(q),z.removeChild(q.domRef),R(q)},u=function(q,z,H){if(q.type==="vtext"){if(q.text!==z.text)q.domRef.textContent=z.text;z.domRef=q.domRef;return}if(q.tag===z.tag&&z.key===q.key&&z["data-component-id"]===q["data-component-id"])z.domRef=q.domRef,T(q,z);else O(q,z,H)},R=function(q){p(q);for(var z in q.children)R(q.children[z])},p=function(q){if(q.onDestroyed)q.onDestroyed()},j=function(q){if(q.onBeforeDestroyed)q.onBeforeDestroyed();if(q.type==="vcomp")q.unmount(q.domRef)},E=function(q){j(q);for(var z in q.children)E(q.children[z])},h=function(q){if(q.onCreated)q.onCreated();if(q.type==="vcomp")i(q)},T=function(q,z){if(!q)q={props:null,css:null,children:[]};if(c(q.props,z.props,z.domRef,z.ns==="svg"),b(q.css,z.css,z.domRef),z.type==="vcomp")return;l(q.children,z.children,z.domRef)},c=function(q,z,H,G){var J;for(var M in q)if(J=z[M],J===void 0)if(G||!(M in H))H.removeAttribute(M,q[M]);else H[M]="";else{if(J===q[M]&&M!=="checked"&&M!=="value")continue;if(G)if(M==="href")H.setAttributeNS("http://www.w3.org/1999/xlink","href",J);else H.setAttribute(M,J);else if(M in H&&!(M==="list"||M==="form"))H[M]=J;else H.setAttribute(M,J)}for(var Q in z){if(q&&q[Q])continue;if(J=z[Q],G)if(Q==="href")H.setAttributeNS("http://www.w3.org/1999/xlink","href",J);else H.setAttribute(Q,J);else if(Q in H&&!(Q==="list"||Q==="form"))H[Q]=z[Q];else H.setAttribute(Q,J)}},b=function(q,z,H){var G;for(var J in q)if(G=z[J],!G)H.style[J]="";else if(G!==q[J])H.style[J]=G;for(var M in z){if(q&&q[M])continue;H.style[M]=z[M]}},a=function(q,z){return q.length>0&&z.length>0&&q[0].key!=null&&z[0].key!=null},l=function(q,z,H){var G=z.length>q.length?z.length:q.length;if(a(z,q))s(q,z,H);else for(var J=0;J<G;J++)X(q[J],z[J],H)},v=function(q){if(q.ns==="svg")q.domRef=document.createElementNS("http://www.w3.org/2000/svg",q.tag);else if(q.ns==="mathml")q.domRef=document.createElementNS("http://www.w3.org/1998/Math/MathML",q.tag);else q.domRef=document.createElement(q.tag)},D=function(q,z){v(q),z(q.domRef),T(null,q),h(q)},i=function(q){var z=q["data-component-id"],H=document.querySelectorAll("[data-component-id='"+z+"']");if(H.length>0){console.error('AlreadyMountedException: Component "'+z+"' is already mounted");return}q.domRef.setAttribute("data-component-id",z),q.mount(function(G){q.children.push(G),q.domRef.appendChild(G.domRef)})},d=function(q,z){if(q.type==="vtext")q.domRef=document.createTextNode(q.text),z.appendChild(q.domRef);else D(q,function(H){z.appendChild(H)})},s=function(q,z,H){var G=0,J=0,M=q.length-1,Q=z.length-1,U,$,_,Y,Z,W,A;for(;;){if(J>Q&&G>M)break;if(U=z[J],$=z[Q],Y=q[G],_=q[M],G>M)X(null,U,H),H.insertBefore(U.domRef,Y?Y.domRef:null),q.splice(J,0,U),J++;else if(J>Q){Z=M;while(M>=G)H.removeChild(q[M--].domRef);q.splice(G,Z-G+1);break}else if(Y.key===U.key)X(q[G++],z[J++],H);else if(_.key===$.key)X(q[M--],z[Q--],H);else if(Y.key===$.key&&U.key===_.key)r(A,_.domRef,Y.domRef,H),o(q,G,M),X(q[G++],z[J++],H),X(q[M--],z[Q--],H);else if(Y.key===$.key)H.insertBefore(Y.domRef,_.domRef.nextSibling),q.splice(M,0,q.splice(G,1)[0]),X(q[M--],z[Q--],H);else if(_.key===U.key)H.insertBefore(_.domRef,Y.domRef),q.splice(G,0,q.splice(M,1)[0]),X(q[G++],U,H),J++;else{W=!1,Z=G;while(Z<=M){if(q[Z].key===U.key){W=!0,A=q[Z];break}Z++}if(W)q.splice(G,0,q.splice(Z,1)[0]),X(q[G++],U,H),H.insertBefore(A.domRef,q[G].domRef),J++;else D(U,function(y){H.insertBefore(y,Y.domRef)}),q.splice(G++,0,U),J++,M++}}},r=function(q,z,H,G){q=z.nextSibling,G.insertBefore(z,H),G.insertBefore(H,q)},o=function(q,z,H){var G=q[z];q[z]=q[H],q[H]=G};function P(q,z,H,G){for(var J in z)q.addEventListener(z[J][0],function(M){F(M,q,H,G)},z[J][1])}var F=function(q,z,H,G){H(function(J){if(q.target)C(q,J,n(z,q.target),[],G)})};function w(q,z,H,G){for(var J in z)q.removeEventListener(z[J][0],function(M){F(M,q,H,G)},z[J][1])}var C=function(q,z,H,G,J){if(!H.length){if(J)console.warn('Event "'+q.type+'" did not find an event handler to dispatch on',z,q);return}else if(H.length>1){G.unshift(z);for(var M=0;M<z.children.length;M++){if(z.type==="vcomp")continue;if(z.children[M].domRef===H[1]){C(q,z.children[M],H.slice(1),G,J);break}}}else{var Q=z.events[q.type];if(Q){var U=Q.options;if(U.preventDefault)q.preventDefault();if(Q.runEvent(q),!U.stopPropagation)S(G,q)}else S(G,q)}},n=function(q,z){var H=[];while(q!==z)H.unshift(z),z=z.parentNode;return H},S=function(q,z){for(var H=0;H<q.length;H++)if(q[H].events[z.type]){var G=q[H].events[z.type],J=G.options;if(J.preventDefault)z.preventDefault();if(G.runEvent(z),J.stopPropagation){z.stopPropagation();break}}};function K(q,z){if(typeof q[0]=="object"){var H=[];for(var G=0;G<q.length;G++)H.push(K(q[G],z));return H}for(G in q)z=z[q[G]];var J;if(z instanceof Array||"length"in z&&z.localName!=="select"){J=[];for(G=0;G<z.length;G++)J.push(K([],z[G]));return J}J={};for(G in t(z)){if(z.localName==="input"&&(G==="selectionDirection"||G==="selectionStart"||G==="selectionEnd"))continue;if(typeof z[G]=="string"||typeof z[G]=="number"||typeof z[G]=="boolean")J[G]=z[G]}return J}var t=function(q){var z={},H=0;do{var G=Object.getOwnPropertyNames(q);for(H=0;H<G.length;H++)z[G[H]]=null}while(q=Object.getPrototypeOf(q));return z},e=function(q){var z=0,H=q.length>0?[q[0]]:[];for(var G=1;G<q.length;G++){if(H[z].type==="vtext"&&q[G].type==="vtext"){H[z].text+=q[G].text;continue}H[++z]=q[G]}return H};function L(q,z,H){var G=0,J;if(!z)if(document.body.childNodes.length>0)J=document.body.firstChild;else J=document.body.appendChild(document.createElement("div"));else if(z.childNodes.length===0)J=z.appendChild(document.createElement("div"));else{while(z.childNodes[G]&&(z.childNodes[G].nodeType===3||z.childNodes[G].localName==="script"))G++;if(!z.childNodes[G])J=document.body.appendChild(document.createElement("div"));else J=z.childNodes[G]}if(!V(q,H,J)){if(q)console.warn("Could not copy DOM into virtual DOM, falling back to diff");while(J.firstChild)J.removeChild(J.lastChild);return H.domRef=J,T(null,H),!1}else if(q){var M=B(!0,H);if(!M)console.warn("Integrity check completed with errors");else console.info("Successfully prerendered page")}return!0}var N=function(q,z,H){if(q)console.warn("VTree differed from node",z,H)},f=function(q){if(q.substr(0,1)=="#"){var z=(q.length-1)/3,H=[17,1,0.062272][z-1];return[Math.round(parseInt(q.substr(1,z),16)*H),Math.round(parseInt(q.substr(1+z,z),16)*H),Math.round(parseInt(q.substr(1+2*z,z),16)*H)]}else return q.split("(")[1].split(")")[0].split(",").map((G)=>{return+G})};function B(q,z){if(z.type=="vtext"){if(z.domRef.nodeType!==3)console.warn("VText domRef not a TEXT_NODE",z),q=!1;else if(z.text!==z.domRef.textContent)console.warn("VText node content differs",z),q=!1}else{if(z.tag.toUpperCase()!==z.domRef.tagName)console.warn("Integrity check failed, tags differ",z.tag.toUpperCase(),z.domRef.tagName),q=!1;if("children"in z&&z.children.length!==z.domRef.childNodes.length)console.warn("Integrity check failed, children lengths differ",z,z.children,z.domRef.childNodes),q=!1;var H=Object.keys(z.props).length,G=null;for(var J=0;J<H;J++)if(G=Object.keys(z.props)[J],G==="href"){var M=window.location.origin+"/"+z.props[G],Q=z.domRef[G],U=z.props[G];if(M!==Q&&U!==Q&&U+"/"!==Q&&M+"/"!==Q)console.warn("Property "+G+" differs",z.props[G],z.domRef[G]),q=!1}else if(G==="height"||G==="width"){if(parseFloat(z.props[G])!==parseFloat(z.domRef[G]))console.warn("Property "+G+" differs",z.props[G],z.domRef[G]),q=!1}else if(G==="class"||G==="className"){if(z.props[G]!==z.domRef.className)console.warn("Property class differs",z.props[G],z.domRef.className),q=!1}else if(!z.domRef[G]){if(z.props[G]!==z.domRef.getAttribute(G))console.warn("Property "+G+" differs",z.props[G],z.domRef.getAttribute(G)),q=!1}else if(z.props[G]!==z.domRef[G])console.warn("Property "+G+" differs",z.props[G],z.domRef[G]),q=!1;H=Object.keys(z.css).length;for(J=0;J<H;J++)if(G=Object.keys(z.css)[J],G==="color"){if(f(z.domRef.style[G]).toString()!==f(z.css[G]).toString())console.warn("Style "+G+" differs",z.css[G],z.domRef.style[G]),q=!1}else if(z.css[G]!==z.domRef.style[G])console.warn("Style "+G+" differs",z.css[G],z.domRef.style[G]),q=!1;for(J=0;J<z.children.length;J++)q&=B(q,z.children[J])}return q}var V=function(q,z,H){var G,J;z.domRef=H,h(z),z.children=e(z.children);for(var M=0;M<z.children.length;M++){if(G=z.children[M],J=H.childNodes[M],!J)return N(q,G,J),!1;if(G.type==="vtext"){if(J.nodeType!==3)return N(q,G,J),!1;if(G.text===J.textContent)G.domRef=J;else return N(q,G,J),!1}else if(G.type==="vcomp")G.mount(function(Q){G.children.push(Q),V(q,G,J)});else{if(J.nodeType!==1)return!1;G.domRef=J,V(q,G,J)}}return!0};function I(q,z){var H=function(){var G=document.getElementById(q);if(G&&G.focus)G.focus()};z>0?setTimeout(H,z):H()}function g(q,z){var H=function(){var G=document.getElementById(q);if(G&&G.blur)G.blur()};z>0?setTimeout(H,z):H()}function x(q){document.body.setAttribute("data-component-id",q)}var m="1.9.0.0";miso={};miso.diff=X;miso.hydrate=L;miso.version=m;miso.delegate=P;miso.callBlur=g;miso.callFocus=I;miso.eventJSON=K;miso.undelegate=w;miso.integrityCheck=B;miso.setBodyComponent=x;
+function X(B,J,z){if(!B&&!J)return;else if(!B&&J)d(J,z);else if(!J)y(B,z);else if(B.type===J.type)u(B,J,z);else h(B,J,z)}var h=function(B,J,z){if(V(B),J.type==="vtext")J.domRef=document.createTextNode(J.text),z.replaceChild(J.domRef,B.domRef);else E(J,function(q){z.replaceChild(q,B.domRef)});W(B)},y=function(B,J){V(B),J.removeChild(B.domRef),W(B)},u=function(B,J,z){if(B.type==="vtext"){if(B.text!==J.text)B.domRef.textContent=J.text;J.domRef=B.domRef;return}if(B.tag===J.tag&&J.key===B.key&&J["data-component-id"]===B["data-component-id"])J.domRef=B.domRef,g(B,J);else h(B,J,z)},W=function(B){k(B);for(var J in B.children)W(B.children[J])},k=function(B){if(B.onDestroyed)B.onDestroyed()},p=function(B){if(B.onBeforeDestroyed)B.onBeforeDestroyed();if(B.type==="vcomp")B.unmount(B.domRef)},V=function(B){p(B);for(var J in B.children)V(B.children[J])},c=function(B){if(B.onCreated)B.onCreated();if(B.type==="vcomp")v(B)},g=function(B,J){if(!B)B={props:null,css:null,children:[]};if(j(B.props,J.props,J.domRef,J.ns==="svg"),b(B.css,J.css,J.domRef),J.type==="vcomp")return;a(B.children,J.children,J.domRef)},j=function(B,J,z,q){var G;for(var H in B)if(G=J[H],G===void 0)if(q||!(H in z))z.removeAttribute(H,B[H]);else z[H]="";else{if(G===B[H]&&H!=="checked"&&H!=="value")continue;if(q)if(H==="href")z.setAttributeNS("http://www.w3.org/1999/xlink","href",G);else z.setAttribute(H,G);else if(H in z&&!(H==="list"||H==="form"))z[H]=G;else z.setAttribute(H,G)}for(var Q in J){if(B&&B[Q])continue;if(G=J[Q],q)if(Q==="href")z.setAttributeNS("http://www.w3.org/1999/xlink","href",G);else z.setAttribute(Q,G);else if(Q in z&&!(Q==="list"||Q==="form"))z[Q]=J[Q];else z.setAttribute(Q,G)}},b=function(B,J,z){var q;for(var G in B)if(q=J[G],!q)z.style[G]="";else if(q!==B[G])z.style[G]=q;for(var H in J){if(B&&B[H])continue;z.style[H]=J[H]}},l=function(B,J){return B.length>0&&J.length>0&&B[0].key!=null&&J[0].key!=null},a=function(B,J,z){var q=J.length>B.length?J.length:B.length;if(l(J,B))s(B,J,z);else for(var G=0;G<q;G++)X(B[G],J[G],z)},i=function(B){if(B.ns==="svg")B.domRef=document.createElementNS("http://www.w3.org/2000/svg",B.tag);else if(B.ns==="mathml")B.domRef=document.createElementNS("http://www.w3.org/1998/Math/MathML",B.tag);else B.domRef=document.createElement(B.tag)},E=function(B,J){i(B),J(B.domRef),g(null,B),c(B)},v=function(B){var J=B["data-component-id"],z=document.querySelectorAll("[data-component-id='"+J+"']");if(z.length>0){console.error('AlreadyMountedException: Component "'+J+"' is already mounted");return}B.domRef.setAttribute("data-component-id",J),B.mount(function(q){B.children.push(q),B.domRef.appendChild(q.domRef)})},d=function(B,J){if(B.type==="vtext")B.domRef=document.createTextNode(B.text),J.appendChild(B.domRef);else E(B,function(z){J.appendChild(z)})},s=function(B,J,z){var q=0,G=0,H=B.length-1,Q=J.length-1,U,A,_,Y,Z,N,M;for(;;){if(G>Q&&q>H)break;if(U=J[G],A=J[Q],Y=B[q],_=B[H],q>H)X(null,U,z),z.insertBefore(U.domRef,Y?Y.domRef:null),B.splice(G,0,U),G++;else if(G>Q){Z=H;while(H>=q)z.removeChild(B[H--].domRef);B.splice(q,Z-q+1);break}else if(Y.key===U.key)X(B[q++],J[G++],z);else if(_.key===A.key)X(B[H--],J[Q--],z);else if(Y.key===A.key&&U.key===_.key)r(M,_.domRef,Y.domRef,z),o(B,q,H),X(B[q++],J[G++],z),X(B[H--],J[Q--],z);else if(Y.key===A.key)z.insertBefore(Y.domRef,_.domRef.nextSibling),B.splice(H,0,B.splice(q,1)[0]),X(B[H--],J[Q--],z);else if(_.key===U.key)z.insertBefore(_.domRef,Y.domRef),B.splice(q,0,B.splice(H,1)[0]),X(B[q++],U,z),G++;else{N=!1,Z=q;while(Z<=H){if(B[Z].key===U.key){N=!0,M=B[Z];break}Z++}if(N)B.splice(q,0,B.splice(Z,1)[0]),X(B[q++],U,z),z.insertBefore(M.domRef,B[q].domRef),G++;else E(U,function(x){z.insertBefore(x,Y.domRef)}),B.splice(q++,0,U),G++,H++}}},r=function(B,J,z,q){B=J.nextSibling,q.insertBefore(J,z),q.insertBefore(z,B)},o=function(B,J,z){var q=B[J];B[J]=B[z],B[z]=q};function T(B,J,z,q){for(var G in J)B.addEventListener(J[G][0],function(H){L(H,B,z,q)},J[G][1])}var L=function(B,J,z,q){z(function(G){if(B.target)w(B,G,n(J,B.target),[],q)})};function R(B,J,z,q){for(var G in J)B.removeEventListener(J[G][0],function(H){L(H,B,z,q)},J[G][1])}var w=function(B,J,z,q,G){if(!z.length){if(G)console.warn('Event "'+B.type+'" did not find an event handler to dispatch on',J,B);return}else if(z.length>1){q.unshift(J);for(var H=0;H<J.children.length;H++){if(J.type==="vcomp")continue;if(J.children[H].domRef===z[1]){w(B,J.children[H],z.slice(1),q,G);break}}}else{var Q=J.events[B.type];if(Q){var U=Q.options;if(U.preventDefault)B.preventDefault();if(Q.runEvent(B),!U.stopPropagation)I(q,B)}else I(q,B)}},n=function(B,J){var z=[];while(B!==J)z.unshift(J),J=J.parentNode;return z},I=function(B,J){for(var z=0;z<B.length;z++)if(B[z].events[J.type]){var q=B[z].events[J.type],G=q.options;if(G.preventDefault)J.preventDefault();if(q.runEvent(J),G.stopPropagation){J.stopPropagation();break}}};function $(B,J){if(typeof B[0]=="object"){var z=[];for(var q=0;q<B.length;q++)z.push($(B[q],J));return z}for(q in B)J=J[B[q]];var G;if(J instanceof Array||"length"in J&&J.localName!=="select"){G=[];for(q=0;q<J.length;q++)G.push($([],J[q]));return G}G={};for(q in t(J)){if(J.localName==="input"&&(q==="selectionDirection"||q==="selectionStart"||q==="selectionEnd"))continue;if(typeof J[q]=="string"||typeof J[q]=="number"||typeof J[q]=="boolean")G[q]=J[q]}return G}var t=function(B){var J={},z=0;do{var q=Object.getOwnPropertyNames(B);for(z=0;z<q.length;z++)J[q[z]]=null}while(B=Object.getPrototypeOf(B));return J};var e=function(B){var J=0,z=B.length>0?[B[0]]:[];for(var q=1;q<B.length;q++){if(z[J].type==="vtext"&&B[q].type==="vtext"){z[J].text+=B[q].text;continue}z[++J]=B[q]}return z};function f(B,J,z){var q=0,G;if(!J)if(document.body.childNodes.length>0)G=document.body.firstChild;else G=document.body.appendChild(document.createElement("div"));else if(J.childNodes.length===0)G=J.appendChild(document.createElement("div"));else{while(J.childNodes[q]&&(J.childNodes[q].nodeType===3||J.childNodes[q].localName==="script"))q++;if(!J.childNodes[q])G=document.body.appendChild(document.createElement("div"));else G=J.childNodes[q]}if(!S(B,z,G)){if(B)console.warn("Could not copy DOM into virtual DOM, falling back to diff");while(G.firstChild)G.removeChild(G.lastChild);return z.domRef=G,populate(null,z),!1}else if(B){var H=K(!0,z);if(!H)console.warn("Integrity check completed with errors");else console.info("Successfully prerendered page")}return!0}var D=function(B,J,z){if(B)console.warn("VTree differed from node",J,z)},m=function(B){if(B.substr(0,1)=="#"){var J=(B.length-1)/3,z=[17,1,0.062272][J-1];return[Math.round(parseInt(B.substr(1,J),16)*z),Math.round(parseInt(B.substr(1+J,J),16)*z),Math.round(parseInt(B.substr(1+2*J,J),16)*z)]}else return B.split("(")[1].split(")")[0].split(",").map((q)=>{return+q})};function K(B,J){if(J.type=="vtext"){if(J.domRef.nodeType!==3)console.warn("VText domRef not a TEXT_NODE",J),B=!1;else if(J.text!==J.domRef.textContent)console.warn("VText node content differs",J),B=!1}else{if(J.tag.toUpperCase()!==J.domRef.tagName)console.warn("Integrity check failed, tags differ",J.tag.toUpperCase(),J.domRef.tagName),B=!1;if("children"in J&&J.children.length!==J.domRef.childNodes.length)console.warn("Integrity check failed, children lengths differ",J,J.children,J.domRef.childNodes),B=!1;var z=Object.keys(J.props).length,q=null;for(var G=0;G<z;G++)if(q=Object.keys(J.props)[G],q==="href"){var H=window.location.origin+"/"+J.props[q],Q=J.domRef[q],U=J.props[q];if(H!==Q&&U!==Q&&U+"/"!==Q&&H+"/"!==Q)console.warn("Property "+q+" differs",J.props[q],J.domRef[q]),B=!1}else if(q==="height"||q==="width"){if(parseFloat(J.props[q])!==parseFloat(J.domRef[q]))console.warn("Property "+q+" differs",J.props[q],J.domRef[q]),B=!1}else if(q==="class"||q==="className"){if(J.props[q]!==J.domRef.className)console.warn("Property class differs",J.props[q],J.domRef.className),B=!1}else if(!J.domRef[q]){if(J.props[q]!==J.domRef.getAttribute(q))console.warn("Property "+q+" differs",J.props[q],J.domRef.getAttribute(q)),B=!1}else if(J.props[q]!==J.domRef[q])console.warn("Property "+q+" differs",J.props[q],J.domRef[q]),B=!1;z=Object.keys(J.css).length;for(G=0;G<z;G++)if(q=Object.keys(J.css)[G],q==="color"){if(m(J.domRef.style[q]).toString()!==m(J.css[q]).toString())console.warn("Style "+q+" differs",J.css[q],J.domRef.style[q]),B=!1}else if(J.css[q]!==J.domRef.style[q])console.warn("Style "+q+" differs",J.css[q],J.domRef.style[q]),B=!1;for(G=0;G<J.children.length;G++)B&=K(B,J.children[G])}return B}var S=function(B,J,z){var q,G;J.domRef=z,callCreated(J),J.children=e(J.children);for(var H=0;H<J.children.length;H++){if(q=J.children[H],G=z.childNodes[H],!G)return D(B,q,G),!1;if(q.type==="vtext"){if(G.nodeType!==3)return D(B,q,G),!1;if(q.text===G.textContent)q.domRef=G;else return D(B,q,G),!1}else if(q.type==="vcomp")q.mount(function(Q){q.children.push(Q),S(B,q,G)});else{if(G.nodeType!==1)return!1;q.domRef=G,S(B,q,G)}}return!0};function O(B,J){var z=function(){var q=document.getElementById(B);if(q&&q.focus)q.focus()};J>0?setTimeout(z,J):z()}function F(B,J){var z=function(){var q=document.getElementById(B);if(q&&q.blur)q.blur()};J>0?setTimeout(z,J):z()}function P(B){document.body.setAttribute("data-component-id",B)}var C="1.9.0.0";miso={};miso.diff=X;miso.hydrate=f;miso.version=C;miso.delegate=T;miso.callBlur=F;miso.callFocus=O;miso.eventJSON=$;miso.undelegate=R;miso.integrityCheck=K;miso.setBodyComponent=P;
